registration form

+++ base/views.py
@@ -1,11 +1,14 @@
 # Create your views here.
 
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .forms import RoomForm
 
 # rooms = [
@@ -15,6 +18,11 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == "POST":
         username = request.POST.get('username')
@@ -33,12 +41,17 @@ def loginPage(request):
         else:
             messages.error(request, 'User OR password does not exist.')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    
+    form = UserCreationForm()
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -62,6 +75,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
 
@@ -73,10 +87,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse('Not allowed')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -86,8 +104,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Not allowed')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
